+++ packages/dart_mappable/test/copy_with/copy_with_subtype_test.dart
@@ -1,6 +1,8 @@
 import 'package:dart_mappable/dart_mappable.dart';
 import 'package:test/test.dart';
 
+import 'copy_with_test.dart';
+
 part 'copy_with_subtype_test.mapper.dart';
 
 @MappableClass(
@@ -49,6 +51,12 @@ class B with BMappable {
   B(this.list, this.a);
 }
 
+@MappableClass(generateMethods: GenerateMethods.copy)
+class ItemContainer2<T> extends ItemContainer<T>
+    with ItemContainer2Mappable<T> {
+  ItemContainer2(super.items);
+}
+
 void main() {
   test('Copies subtype', () {
     var zoo = Zoo(Cat('Mauzi', 'Black'));
@@ -60,5 +68,19 @@ void main() {
     var z3 = zoo.copyWith.animal.$update((a) => Dog(a.name));
     expect(z3.animal, isA<Dog>());
     expect(z3.animal.name, equals('Mauzi'));
+  });
+
+  test('Copies imported subtype', () {
+    var container = ItemContainer2([
+      BrandList([Brand('Audi')]),
+    ]);
+    var c2 = container.copyWith.items.at(0).items.add(Brand('BMW'));
+    expect(c2.items[0], isA<BrandList>());
+    expect(c2.items[0].items.last, equals(Brand('BMW')));
+
+    expect(identical(container.items, c2.items), isFalse);
+    expect(identical(container.items[0], c2.items[0]), isFalse);
+    expect(container.items[0].items, hasLength(1));
+    expect(c2.items[0].items, hasLength(2));
   });
 }