+++ packages/dart_mappable/test/serialization/basic_serialization_test.mapper.dart
@@ -2,6 +2,7 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // dart format off
 // ignore_for_file: type=lint
+// ignore_for_file: invalid_use_of_protected_member
 // ignore_for_file: unused_element, unnecessary_cast, override_on_non_overriding_member
 // ignore_for_file: strict_raw_type, inference_failure_on_untyped_parameter
 
@@ -75,7 +76,7 @@ class AMapper extends ClassMapperBase<A> {
   static String _$a(A v) => v.a;
   static const Field<A, String> _f$a = Field('a', _$a);
   static int _$b(A v) => v.b;
-  static const Field<A, int> _f$b = Field('b', _$b, opt: true, def: 0);
+  static const Field<A, int> _f$b = Field('b', _$b, opt: true, def: A.defaultB);
   static double? _$c(A v) => v.c;
   static const Field<A, double> _f$c = Field('c', _$c, opt: true);
   static bool _$d(A v) => v.d;